==> rki_vorhersagen/01_rki_data_preperation.ipynb <==
# <span style="color:#f37726">01 RKI Data Preperation</span>

**Ziele des Notebooks:**
- Attribute selektieren
- Datentypen anpassen und fehlende Werte ersetzen
- Landkreisdaten auslagern
- Gemeldete Fälle komprimieren
- Impffortschritt hinzufügen

# Benötigte Bibliotheken für dieses Notebook
import numpy as np
import pandas as pd

## Selektion der Attribute
Es soll entschieden werden, welche der 18 Attribute für das Vorhersagemodell relevant sind:

**FID** *≙ Index in der csv-Datei*
<br>Irrelevant. Primärschlüssel, jedoch ohne Bedeutung für die Vorhersage. Wird vom pandas DataFrame Index abgelöst.

**IdBundesland** *≙ Id des Bundeslands*	
Irrelevant. Name des Bundeslandes vorhanden.

**Bundesland** *≙ Name des Bundeslandes*
<br> Relevant. Ermöglicht geograhpische Modellierung.

**Landkreis** *≙ Name des Landkreises*
<br>Relevant. Ermöglicht geographische Modellierung

**Altersgruppe** *≙ Altersgruppe des Falles aus 6 Gruppen*
<br>Irrelevant. Fallzahlen sollen altersunabhänig vorhergesagt werden. Die Alterstruktur der Landkreise wird als demographischer Faktor im Clustering berücksichtigt.

**Geschlecht** *≙ Geschlecht des Falle*
<br>Irrelevant. Fallzahlen sollen unabhängig vom Geschlecht vorhergesagt werden.

**AnzahlFall** *≙ Anzahl der Fälle in der entsprechenden Gruppe*
<br>Relevant. Wird zu Fallzahlen akkumuliert.

**AnzahlTodesfall** *≙ Anzahl der Todesfälle in der entsprechenden Gruppe*
<br>Relevant. Wird (eventuell) zur Modellierung der infektionsgefährdeten Bevölkerung verwendet.

**Meldedatum** *≙ Datum, wann der Fall dem Gesundheitsamt bekannt wurde*
<br>Relevant. Ermöglicht zeitliche Modellierung.

**IdLandkreis** *≙ Id des Landkreises*
<br>Relevant. Erleichtert das Verbinden der Datensätze.

**Datenstand** *≙ Datum, wann der Datensatz zuletzt aktualisiert*
<br>Irrelevant. Der Datensatz wird täglich aktualisiert. Wir arbeiten mit der Version vom 13. April 2021.

**IdLandkreis** *≙ Id des Landkreises*
<br>Relevant. Erleichtert das Verbinden der Datensätze. Wird ebenfalls vom Statistischen Bundesamt verwendet.

**NeuerFall** *≙ Fall ist in Publikation des Vortages enthalten*
<br>Irrelevant. Meldefehler haben nur Einfluss auf dynamische Analysen (Echtzeit).

**NeuerTodesfall** *≙ Todesfall ist in Publikation des Vortages enthalten*
<br>Irrelevant. Meldefehler haben nur Einfluss auf dynamische Analysen (Echtzeit).

**Refdatum** *≙ Erkrankungsdatum*
<br>Irrelevant. Das Meldedatum wird zur zeitlichen Modellierung verwendet.

**NeuGenesen** *≙ Genesung ist in Publikation des Vortages enthalten*
<br>Irrelevant. Meldefehler haben nur Einfluss auf dynamische Analysen (Echtzeit).

**AnzahlGenesen** *≙ Anzahl der Genesenen in der entsprechenden Gruppe*
<br>Relevant. Wird (eventuell) zur Modellierung der infektionsgefährdeten Bevölkerung verwendet.

**IstErkrankungsbeginn** *≙ Ist Refdatum gleich Erkrankungsbeginn*
<br>Irrelevant. Das Meldedatum wird zur zeitlichen Modellierung verwendet.

**Altersgruppe2** *≙ Altersgruppe des Falles aus 5-Jahresgruppen*
<br>Irrelevant. Fallzahlen sollen altersunabhänig vorhergesagt werden. Die Alterstruktur der Landkreise wird als demographischer Faktor im Clustering berücksichtigt.

# Laden des Datensatzes als pandas DataFrame
data = pd.read_csv(r'data/RKI_COVID19.csv')
df = pd.DataFrame(data)

shape0 = df.shape[1]

# Entferne Spalten mit irrelevanten Attributen
df.drop("FID", axis=1, inplace=True)
df.drop("IdBundesland", axis=1, inplace=True)
df.drop("Altersgruppe", axis=1, inplace=True)
df.drop("Geschlecht", axis=1, inplace=True)
df.drop("Datenstand", axis=1, inplace=True)
df.drop("Refdatum", axis=1, inplace=True)
df.drop("IstErkrankungsbeginn", axis=1, inplace=True)
df.drop("Altersgruppe2", axis=1, inplace=True)
df.drop("NeuerFall", axis=1, inplace=True)
df.drop("NeuerTodesfall", axis=1, inplace=True)
df.drop("NeuGenesen", axis=1, inplace=True)

shape1 = df.shape[1]

print(f"Anzahl entfernter Attribute: {str(shape0-shape1)}")

## Datenbereinigung
Fehlende Werte sollen entdeckt und ggf. mit sinnvollen Werten ersetzt werden:

# Erstelle Series mit NaN Zähler
nan_series = df.isna().sum()

# Erstelle Series mit Datentypen
types_series = df.dtypes

# Erstelle Series mit einzigartigen Werten
unique_series = df.nunique()

# Verschmelzen der Serien zu DataFrame
parameter_overview_df = pd.concat([unique_series, nan_series, types_series], axis=1)
parameter_overview_df.columns = ['distinct_values', 'NaN_count', 'dtype']
                                  
print(parameter_overview_df)
print("\n")
print("Form des DataFrame: " + str(df.shape))

Die Datentypen der Attribute sind bereits kohärent mit den jeweils möglichen Ausprägungen. Lediglich das Meldedatum soll zu 'datetime64' umgewandelt werden. Boolsche Variablen existieren in ursprünglicher Form nicht. Die Attribute 'Landkreis', 'Bundesland' und 'Meldedatum' sind nominal skaliert. Die Attribute 'AnzahlFall', 'AnzahlTodesfall' und 'AnzahlGenesen' metrisch.

df['Meldedatum'] =  pd.to_datetime(df['Meldedatum'], format='%Y/%m/%d %H:%M:%S')
print(df['Meldedatum'].dtype)

## Modifikation der Landkreisdaten

display(df.head())

Die Fallzahlen, die Todesfälle und die Genesungen sollen pro Tag und Landkreis akkumuliert werden. Die Auslagerung von Landkreisnamen und entsprechendem Bundesland in einen weiteren Dataframe erleichtert die Kompression:

landkreise_df = df[['Landkreis', 'Bundesland', 'IdLandkreis']]
landkreise_df = landkreise_df.groupby(['IdLandkreis']).last()
display(landkreise_df)

Die Art des Kreises (Landkreis, kreisfreie Stadt) soll als boolsche Spalte ergänzt werden. Dazu wird eine selbstgeschriebene Funktion verwendet:

from rki_function0 import AddBoolColumn

landkreise_df['KreisArt']=landkreise_df.Landkreis.str[:2]
AddBoolColumn(landkreise_df, 'KreisArt', 'SK')
landkreise_df.drop("KreisArt", axis=1, inplace=True)
landkreise_df.rename(columns={"SK": "IstStadt"}, inplace=True)
landkreise_df['Landkreis'] = landkreise_df.Landkreis.str[3:]
display(landkreise_df)

Gemäß der Beobachtung im ersten Notebook sollen noch die Berliner Stadtbezirke zu einem Landkreis aggregiert werden:

berliner_ids_mask = landkreise_df.Landkreis.str.contains("Berlin")
berliner_ids = berliner_ids_mask.index[berliner_ids_mask]

# Entfernen der 12 Bezirke
landkreise_df.drop(berliner_ids, axis=0, inplace=True)

# Hinzufügen des Berliner Bezirks
landkreise_df.loc[11000] = ['Berlin', 'Berlin', True]
landkreise_df.sort_index(inplace=True)

Außerdem soll der DataFrame die Bevölkerungszahl des jeweiligen Landkreises enthalten. Das Statistische Bundesamt [veröffentlicht](https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales/Gemeindeverzeichnis/Administrativ/04-kreise.html) den Bevölkerungstand jährlich als xlsx-Datei. Für das Einlesen haben wir eine csv-Datei erstellt:

data = pd.read_csv(r'data/landkreise_einwohner_fläche.csv', encoding='cp1252', delimiter=';')
landkreise_pop_df = pd.DataFrame(data)
landkreise_pop_df.set_index('IdLandkreis', drop=True, inplace=True)
display(landkreise_pop_df)
print(landkreise_pop_df.dtypes)

# Überflüssige Spalten entfernen
landkreise_pop_df.drop('Kreisart', axis=1, inplace=True)
landkreise_pop_df.drop('Landkreis', axis=1, inplace=True)
landkreise_pop_df.drop('Bevölkerung_Männlich', axis=1, inplace=True)
landkreise_pop_df.drop('Bevölkerung_Weiblich', axis=1, inplace=True)
landkreise_pop_df.drop('Einwohner_pro_qkm', axis=1, inplace=True)

# Landkreis DataFrame erweitern
landkreise_df = pd.concat([landkreise_df, landkreise_pop_df], axis=1)
display(landkreise_df)

## Kompression der gemeldeten Fälle

Die Fallzahlen sollen für jeden Landkreis als Tageswert vorliegen. Für die Umsetzung bietet sich [pandas gropby-Funktion](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.groupby.html) in Kombination mit einem Multiindex an:

df.drop("Bundesland", axis=1, inplace=True)
df.drop("Landkreis", axis=1, inplace=True)

df = df.groupby(['IdLandkreis', 'Meldedatum']).sum() # Multiindex

display(df)

Der zeitliche Horizont der Meldedaten soll angepasst werden. Beginnend ab März 2020 bis April 2021 werden dann auch Tage ohne gemeldete Fälle im Datensatz ergänzt:

start_time = pd.to_datetime('2020-03-01')
end_time = pd.to_datetime('2021-03-31')
case_dates = pd.date_range(start_time, end_time, freq='D')


df = df.loc[(slice(None), case_dates), :]
df = df.reindex(pd.MultiIndex.from_product([df.index.levels[0], case_dates], names=['IdLandkreis', 'Meldedatum']), fill_value=0)
display(df)

Abschließend sollen auch in diesem DataFrame die Berliner Bezirke zu einem Landkreis aggregiert werden. Dafür wird erneut eine selbstgeschriebene Funktion verwendet:

from rki_function1 import CountyCasesAggregator


df = CountyCasesAggregator(df, berliner_ids, 11000)

print("Berlin hinzugefügt?: " + str(11000 in df.index))
print("\n")
print("Berliner Bezirke entfernt?:")
for id in berliner_ids:
    print(id not in df.index.get_level_values(0).tolist())

In einem weiteren DataFrame wird zu den Kalenderdaten noch der Wochentag und Monat gespeichert:

case_dates_df = pd.DataFrame(case_dates, columns=['Meldedatum'])
case_dates_df['Monat']=case_dates_df['Meldedatum'].dt.month
case_dates_df['Wochentag'] = case_dates_df['Meldedatum'].dt.dayofweek
case_dates_df['Jahr'] = case_dates_df['Meldedatum'].dt.year
case_dates_df.set_index('Meldedatum', drop=True, inplace=True)
display(case_dates_df)

## Impffortschritt hinzufügen

Für die Modellierung der ansteckungsgefährdenden Bevölkerung soll der Impffortschritt berücksichtigt werden. Die Daten dazu stellt das Robert Koch-Institut zusammen mit Visualisierungen in einem [Dashboard](https://impfdashboard.de/) bereit:

data = pd.read_csv(r'data/germany_vaccinations_timeseries_v2.tsv', sep='\t')
vaccines_df = pd.DataFrame(data)
display(vaccines_df)

Wichtig für die Modellierung ist lediglich die bundesweite Quote an Erst- und Vollimpfungen. Zwischen den Impfstoffen soll nicht unterschieden werden. Auch die Priorisierung wird vernachlässigt:

# Entferne Spalten mit irrelevanten Attributen
vaccines_df = vaccines_df[['date', 'impf_quote_erst', 'impf_quote_voll']]
vaccines_df.rename(columns={"date": "Meldedatum"}, inplace=True)
vaccines_df['Meldedatum'] =  pd.to_datetime(vaccines_df['Meldedatum'], format='%Y/%m/%d %H:%M:%S')
vaccines_df.set_index('Meldedatum', drop=True, inplace=True)
display(vaccines_df)

Die Exploration der komprimierten Fallzahlen ist Thema des nächsten Notebooks. Dazu sollen die DataFrames als pickle bzw. csv-Datein exportiert werden:

df.to_pickle(r'data/fallzahlen_akkumuliert.pkl')
df.to_csv(r'data/fallzahlen_akkumuliert.csv', sep=';', encoding='utf-8', index=True)

landkreise_df.to_pickle(r'data/landkreise_bundesland_id.pkl')
landkreise_df.to_csv(r'data/landkreise_bundesland_id.csv', sep=';', encoding='utf-8', index=True)

case_dates_df.to_pickle(r'data/case_dates.pkl')
case_dates_df.to_csv(r'data/case_dates.csv', sep=';', encoding='utf-8', index=True)

vaccines_df.to_pickle(r'data/vaccine_count.pkl')
vaccines_df.to_csv(r'data/vaccine_count.csv', sep=';', encoding='utf-8', index=True)
